2nd commit
Made tabs and added content to classes and home. Also made the database functions and views as described in the config.php file. Removed the extra pages made for each tab and added the content to the home page using the switch statement. Login and Logout are made easy and insert data into the log table for keeping records of user logins.